requirements.txt
asgiref==3.11.1
Django==5.1.6
psycopg2-binary==2.9.11
sqlparse==0.5.5
python-dotenv
django-allauth
requests
PyJWT
cryptography
Pillow==12.1.1
python-frontmatter
PyYAML
markdown
